TL;DR - This old man done fucked up and made shit that be really stupid, yo. Much thanks to @DarkGuardsman for pointing out the idiocy and beating me over the head with it.
When I tried to add the Fluid Discharge we started having the old OOM's and NPE's that I thought I'd squashed.
After bouncing some ideas off @DarkGuardsman it was suggested that I take a look at "Assembly Line" and "Resonant Induction".
In doing so I discovered we could have some helpers in a better place and that the Fluid Discharge was almost acting like "magic" in that it'd discharge fluids into a 32x32x32 area instead of just under it.
At the same time it became clear that the code I had for the drain was nonsensical as it looked for anything within a 32 block distance above it, ignoring things like boundary walls and other separators.

This is the start of a rewrite of both to make more sense prior to moving on to adding the other machines